\name{d.broeder}
\alias{d.broeder}
\docType{data}
\title{
Broeder & Schuetz (2009) Experiment 3
}
\description{
The data from Broeder & Schuetz (2009) Experiment 3, used as an example in MPTinR
}
\references{
Broeder, A., & Schuetz, J. (2009). Recognition ROCs are curvilinear-or are they? On premature arguments against the two-high-threshold model of recognition. \emph{Journal of Experimental Psychology: Learning, Memory, and Cognition}, 35(3), 587. doi:10.1037/a0015279
}
\usage{data(d.broeder)}
\keyword{datasets}
