\name{MPTinR-package}
\alias{MPTinR-package}
\alias{MPTinR}
\docType{package}
\title{
\packageTitle{MPTinR}
}
\description{
\packageDescription{MPTinR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MPTinR}
\packageIndices{MPTinR}

To fit MPT Models use \code{\link{fit.mpt}}, to fit other models use \code{\link{fit.model}} or \code{\link{fit.mptinr}} (which is called by the other two functions).

For model selection use \code{\link{select.mpt}}.

A helper function for writing model files is \code{\link{check.mpt}}


}
\references{
Riefer, D. M., & Batchelder, W. H. (1988). Multinomial modeling and the measurement of cognitive processes. \emph{Psychological Review}, 95, 318-339

Singmann, H., & Kellen, D. (2013). MPTinR: Analysis of multinomial processing tree models in R. \emph{Behavior Research Methods}, 45(2), 560-575. \url{https://doi.org/10.3758/s13428-012-0259-0}
}

