\name{rb.fig1.data}
\alias{rb.fig1.data}
\alias{rb.fig2.data}
\docType{data}
\title{
Data to be used for the examples of MPTinR.
}
\description{
Dataset 1 (fig1) is taken from Riefer & Batchelder (1988, Table 1) and contains multiple individuals.\cr
Dataset 2 (fig2) is taken from Riefer & Batchelder (1988, Table 3).\cr
}
\usage{
data(rb.fig1.data)
data(rb.fig2.data)
}

\source{
Riefer, D. M., & Batchelder, W. H. (1988). Multinomial modeling and the measurement of cognitive processes. \emph{Psychological Review}, 95, 318-339.
}


\keyword{datasets}
